refactor: move back to single file button.rs
