--- src/widgets/button/mod.rs
@@ -1,5 +1,0 @@
-mod button;
-mod theme;
-mod widget;
-
-pub use button::Button;
